:recycle: Refactor: separando formularios por componentes
